Don't crash on exported internals
What the hell is an exported internal? Well, right now it is something
that lezer-generator needs from lezer, but which normal consumers won't
actually need to look at. Similarly, you might need to export something
for testing purposes, but don't want it to show up in the docs.

The concept remains a bit dodgy, but I think it's useful enough to
try and handle it.